Made the changes that were requested.
